@@ -3,11 +3,13 @@ apply plugin: 'com.android.application'
 archivesBaseName = 'android-crop-example'
 
 android {
-    compileSdkVersion 27
+    namespace "com.soundcloud.android.crop.example"
+
+    compileSdkVersion 36
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 27
+        minSdkVersion 21
+        targetSdkVersion 33
         versionCode Integer.parseInt(project.VERSION_CODE)
         versionName project.VERSION
     }
@@ -15,6 +17,5 @@ android {
 
 dependencies {
     implementation project(':lib')
-    implementation 'com.android.support:support-compat:27.1.1'
-    implementation 'com.android.support:support-core-utils:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.7.1'
 }
