@@ -1,35 +1,58 @@
 apply plugin: 'com.android.library'
-apply plugin: 'maven'
+apply plugin: 'maven-publish'
 apply plugin: 'signing'
 //apply from: '../.publishing/sonatype.gradle'
 
 archivesBaseName = 'android-crop'
 
 android {
-    compileSdkVersion 27
+    namespace "com.soundcloud.android.crop"
+
+    compileSdkVersion 36
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 15
-        versionName "1.0.5"
+        minSdkVersion 21
+        targetSdkVersion 33
+        versionCode 16
+        versionName "1.0.6"
 
         testApplicationId 'com.soundcloud.android.crop.test'
         testInstrumentationRunner 'android.test.InstrumentationTestRunner'
     }
-    lintOptions {
+    lint {
         disable 'Registered', 'IconDensities'
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+    }
+    publishing {
+        singleVariant("release")
     }
 }
 
 dependencies {
-    implementation 'com.android.support:support-annotations:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:exifinterface:27.1.1'
-    androidTestImplementation 'com.squareup:fest-android:1.0.7'
-    androidTestImplementation 'com.android.support:support-v4:27.1.1'
-    androidTestImplementation 'org.mockito:mockito-core:1.9.5'
-    androidTestImplementation 'com.google.dexmaker:dexmaker:1.0'
-    androidTestImplementation 'com.google.dexmaker:dexmaker-mockito:1.0'
+    implementation 'androidx.appcompat:appcompat:1.7.1'
+    implementation 'androidx.annotation:annotation:1.9.1'
+    implementation 'androidx.exifinterface:exifinterface:1.4.1'
+
+    androidTestImplementation 'com.squareup:fest-android:1.0.8'
+    androidTestImplementation 'com.android.support:support-v4:28.0.0'
+    androidTestImplementation 'org.mockito:mockito-core:5.18.0'
+    androidTestImplementation 'com.google.dexmaker:dexmaker:1.2'
+    androidTestImplementation 'com.google.dexmaker:dexmaker-mockito:1.2'
 }
 
+publishing {
+    publications {
+        release(MavenPublication) {
+            groupId = 'com.github.yaroslav-v'
+            artifactId = 'android-crop'
+            version = '1.0.6'
+
+            afterEvaluate {
+                from components.release
+            }
+        }
+    }
+}
